pyproject.toml
[project]
name = "scraper-boletin-oficial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4==4.12.3",
    "ipykernel>=6.30.1",
    "openai>=1.106.1",
    "openpyxl>=3.1.0",
    "pandas==2.2.0",
    "pyarrow==15.0.0",
    "python-dotenv>=1.0.0",
    "requests==2.31.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]
